ARM: sunxi: Adapt to the unsigned long change for get_ram_size()
get_ram_size() now properly returns unsigned long. This patch adapts
sunxi to also work with unsigned long's.

+++ board/sunxi/dram_a13_mid.c
@@ -25,7 +25,7 @@ static struct dram_para dram_para = {
 	.emr3 = 0,
 };
 
-int sunxi_dram_init(void)
+unsigned long sunxi_dram_init(void)
 {
 	return dramc_init(&dram_para);
 }
